Created and tested cell line repo
From this commit, the development of the APIs and pipelines will be
halted, until a commit includes these files. I will only be developing
(and fixing) the repo, domain and infraestructure modules.

For this reason (while the content on APIs is ok) the content on the
pipeline module is not to be taken as valid, but as extremely buggy
code.