RM args from sql operators and edit exp_repo
In this commit, I've edited the SQL operation decorators, removing their
parameters. Previously, I could indicate if I wanted to return True if
correct or False if not, or a value if correct and None if not. Now, the
decorators return whatever the wrapped function returns, and raise if
there's an error.

Regarding the ExperimentRepo:
- Added the exp classification and source table creation and
`get_or_create` methods to the ExperimentRepo.
- Removed validation from the `get_or_create` method for experiment and
experiment_score creating.
- Simplified the Experiment dataclass